Fix bug where files are deleted in termux

+++ YTD_Android.py
@@ -3,6 +3,7 @@ import sys
 import json
 import linecache
 import logging
+import platform
 from termcolor import colored
 
 # Constants
@@ -99,7 +100,8 @@ def update_history(title, site):
     if data["default"][0]["history_backup"] == "y":
         sync_with_drive()
 
-    os.remove(TEMP_LOC)
+    if platform.system() == 'Linux' and 'com.termux' not in os.getenv('PREFIX', ''):
+        os.remove(TEMP_LOC)
     logging.info("History updated and temp file removed.")
     sys.exit()
 
@@ -271,8 +273,9 @@ def download_from_others():
     try:
         download_content(opt, site=dir_name)
     except Exception as e:
-        os.rmdir(path)
         logging.error(f"Error downloading from {dir_name}: {e}")
+        if platform.system() == 'Linux' and 'com.termux' not in os.getenv('PREFIX', ''):
+            os.rmdir(path)
 
 def download_from_ftp_or_torrent():
     """Download content from FTP or torrent links"""
@@ -331,7 +334,8 @@ def master_directory():
         """Remove empty directories"""
         empty_dirs = [root for root, dirs, files in os.walk(GEN_PATH) if not dirs and not files]
         for empty_dir in empty_dirs:
-            os.rmdir(empty_dir)
+            if platform.system() == 'Linux' and 'com.termux' not in os.getenv('PREFIX', ''):
+                os.rmdir(empty_dir)
 
     clean_empty_directories()
     link_distributor()
@@ -368,3 +372,7 @@ if __name__ == "__main__":
 
     # Start the master directory process
     master_directory()
+
+    # Delete TEMP_LOC if running on Linux but not Termux
+    if os.path.isfile(TEMP_LOC) and platform.system() == 'Linux' and 'com.termux' not in os.getenv('PREFIX', ''):
+        os.remove(TEMP_LOC)
